feat: login is now rich on fragments on basic navigation
