+++ app/src/main/res/values/strings.xml
@@ -813,4 +813,6 @@
 
     <string name="la_dropdown_add_new">Add new server</string>
     <string name="la_shortcut_label">Open login editor (beta)</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
